Fixed an issues found by clang warnings. In advance_pattern in Channel_MODULE.cc, the portamento flag wasn't being saved properly due an autocomplete error. Eliminated all warnings found by clang. Went back to g++ and eliminated any new warnings that came up.

+++ conversion.cc
@@ -30,7 +30,7 @@ namespace MultiPLAY
 
 	extern unsigned long from_lsb4_lu(unsigned char in[4])
 	{
-		return in[0] | (in[1] << 8) | (in[2] << 16) | (in[3] << 24);
+		return (unsigned long)(in[0] | (in[1] << 8) | (in[2] << 16) | (in[3] << 24));
 	}
 
 	extern short from_msb2(unsigned char in[2])
